deleted unused React logos, consolidated navbar components, replaced navbar logo to have transparent background, changed background of side menu to match navbar
