+++ client/src/stylesheets/Burger.scss
@@ -1,5 +1,5 @@
  $primary-color: rgb(95, 171, 117); 
- $color:  rgb(3, 103, 179);
+ $color:  #1f7bc5;
  $dark-blue: rgb(0, 38, 72);
  $dark-blue-line: rgba(0, 38, 72, 0.5);
  $font: "ITC Avant Garde Gothic",sans-serif;
