+++ client/src/stylesheets/NavBar.scss
@@ -4,11 +4,6 @@ $dark-blue: rgb(0, 38, 72);
 nav {
     background-color: $color;
     margin-bottom: 0px;
-    
-    .logo {
-        height: 100px;
-        width: 100px;
-      }
 
     .cart {
         position: absolute;
